Migrated to a lazy row instead of facets instead of hard coded bubbles.
